requirements.txt
asgiref==3.5.2
Django==3.2.15
django-taggit==3.0.0
image-sorting-tool==0.1.0
Pillow==9.2.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
